#10 New data structure for local data storage
Also there is some possibility for stability.
