+++ app/src/main/java/android/revesz/seriestracker_v2/data/SeriesRepository.kt
@@ -1,19 +1,19 @@
 package android.revesz.seriestracker_v2.data
 
-class SeriesRepository private constructor(private val seriesDao: SeriesDao) {
+class SeriesRepository private constructor(private val localDataDao: LocalDataDao) {
 
-    fun getSeries() = seriesDao.getAll()
+    fun getSeries() = localDataDao.getAll()
 
-    fun getShow(showId: Int) = seriesDao.getItemByID(showId)
+    fun getShow(showId: Int) = localDataDao.getItemByID(showId)
 
     companion object {
 
         // For Singleton instantiation
         @Volatile private var instance: SeriesRepository? = null
 
-        fun getInstance(seriesDao: SeriesDao) =
+        fun getInstance(localDataDao: LocalDataDao) =
             instance ?: synchronized(this) {
-                instance ?: SeriesRepository(seriesDao).also { instance = it }
+                instance ?: SeriesRepository(localDataDao).also { instance = it }
             }
     }
 }
