+++ app/src/main/res/layout/list_item_shows.xml
@@ -1,18 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:orientation="vertical"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    xmlns:tools="http://schemas.android.com/tools">
 
     <data>
-        <variable
-            android:name="clickListener"
-            android:type="android.view.View.OnClickListener"/>
-        <variable
-            android:name="series"
-            android:type="android.revesz.seriestracker_v2.data.Series"/>
+        <variable name="clickListener" type="android.view.View.OnClickListener"/>
+        <variable name="series" type="android.revesz.seriestracker_v2.data.LocalData"/>
     </data>
 
     <android.revesz.seriestracker_v2.views.MaskedCardView
@@ -22,8 +15,8 @@
         android:layout_marginEnd="@dimen/card_side_margin"
         android:layout_marginBottom="@dimen/card_bottom_margin"
         android:onClick="@{clickListener}"
-        android:cardElevation="@dimen/card_elevation"
-        android:cardPreventCornerOverlap="false"
+        app:cardElevation="@dimen/card_elevation"
+        app:cardPreventCornerOverlap="false"
         app:shapeAppearanceOverlay="@style/ShapeAppearance.Series.Card">
 
         <androidx.constraintlayout.widget.ConstraintLayout
@@ -36,7 +29,6 @@
                 android:layout_height="@dimen/show_item_image_height"
                 android:contentDescription="@string/picture_of_show"
                 android:scaleType="centerCrop"
-                app:imageFromUrl="@{series.imageUrl}"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toTopOf="parent" />
@@ -60,4 +52,4 @@
 
     </android.revesz.seriestracker_v2.views.MaskedCardView>
 
-</LinearLayout>
+</layout>
